+++ tt_metal/tt-llk/tests/python_tests/helpers/param_config.py
@@ -12,6 +12,14 @@ from helpers.tile_shape import construct_tile_shape
 from typing_extensions import deprecated
 
 from .chip_architecture import ChipArchitecture, get_chip_architecture
+from .constraints import (
+    _quasar_effective_sfpu_format,
+    _quasar_fpu_source_format,
+    _quasar_sfpu_hardware_format,
+    is_valid_quasar_fpu_path,
+    is_valid_quasar_packer_conversion,
+    is_valid_quasar_unpack_to_dest,
+)
 from .data_format_inference import is_format_combination_outlier
 from .format_config import (
     DataFormat,
@@ -478,7 +486,6 @@ class QuasarSfpuVariant:
     sfpu_src: DataFormat
     sfpu_dst: DataFormat
     pack_src: DataFormat
-    int32_dest: bool = False
 
     @property
     def uses_fpu(self) -> bool:
@@ -495,197 +502,6 @@ class QuasarSfpuVariant:
             config.sfpu_dst = self.sfpu_dst
             config.pack_src = self.pack_src
             config.pack_S_src = self.pack_src
-
-
-# Quasar conversion capabilities for L1 -> registers, SrcA -> Dest, and
-# Dest -> L1. Sources: Tensix unpacker/packer conversions and Neo FPU formats.
-_QUASAR_NARROW_FORMATS = frozenset(
-    {
-        DataFormat.Float16,
-        DataFormat.Float16_b,
-        DataFormat.Fp8_e4m3,
-        DataFormat.MxFp8R,
-        DataFormat.MxFp8P,
-        DataFormat.MxFp4,
-        DataFormat.MxInt8,
-        DataFormat.MxInt4,
-        DataFormat.MxInt2,
-    }
-)
-_QUASAR_NARROW_FLOAT_DEST_FORMATS = frozenset(
-    {DataFormat.Float16, DataFormat.Float16_b}
-)
-_QUASAR_UNPACK_TO_DEST_FORMATS = {
-    # TF32 is stored in a 32-bit Float32 container in Dest.
-    DataFormat.Float32: {DataFormat.Float32} | _QUASAR_NARROW_FLOAT_DEST_FORMATS,
-    DataFormat.Tf32: {DataFormat.Float32} | _QUASAR_NARROW_FLOAT_DEST_FORMATS,
-    **{
-        input_format: _QUASAR_NARROW_FLOAT_DEST_FORMATS
-        for input_format in _QUASAR_NARROW_FORMATS
-    },
-    DataFormat.Int32: {DataFormat.Int32},
-    DataFormat.Int16: {DataFormat.Int16},
-    DataFormat.Int8: {DataFormat.Int8},
-    DataFormat.UInt8: {DataFormat.UInt8},
-}
-
-# SrcA/B additionally permit a TF32 result for narrow floating-point inputs.
-_QUASAR_UNPACK_TO_SRCA_FORMATS = {
-    **{
-        input_format: output_formats
-        for input_format, output_formats in _QUASAR_UNPACK_TO_DEST_FORMATS.items()
-        # Int32 may be unpacked only to Dest or SrcS, not SrcA/B.
-        if input_format != DataFormat.Int32
-    },
-    DataFormat.Float32: {DataFormat.Tf32} | _QUASAR_NARROW_FLOAT_DEST_FORMATS,
-    DataFormat.Tf32: {DataFormat.Tf32} | _QUASAR_NARROW_FLOAT_DEST_FORMATS,
-    **{
-        input_format: {DataFormat.Tf32} | _QUASAR_NARROW_FLOAT_DEST_FORMATS
-        for input_format in _QUASAR_NARROW_FORMATS
-    },
-    # Quasar-only 2x-packed register formats are legal only for MXFP4 -> SrcA/B.
-    DataFormat.MxFp4: {
-        DataFormat.Tf32,
-        DataFormat.Float16,
-        DataFormat.Float16_b,
-        DataFormat.MxFp4_2x_A,
-        DataFormat.MxFp4_2x_B,
-    },
-}
-
-# Effective SrcA -> Dest conversions supported by Quasar unary datacopy. The
-# LLK selects MOVA2D/MOVB2D for narrow Dest and ELWADD for 32-bit Dest; test
-# generation only needs to know whether the complete datacopy path is executable.
-_QUASAR_FPU_DATACOPY_DEST_FORMATS = {
-    DataFormat.Float16: {DataFormat.Float16, DataFormat.Float32},
-    DataFormat.Float16_b: {DataFormat.Float16_b, DataFormat.Float32},
-    DataFormat.Tf32: {DataFormat.Float32},
-    DataFormat.Int8: {DataFormat.Int8, DataFormat.Int32},
-    DataFormat.UInt8: {DataFormat.UInt8, DataFormat.Int32},
-    DataFormat.Int16: {DataFormat.Int16},
-}
-
-_QUASAR_PACK_OUTPUTS = {
-    DataFormat.Float32: {
-        DataFormat.Float32,
-        DataFormat.Tf32,
-    }
-    | _QUASAR_NARROW_FORMATS,
-    DataFormat.Float16: _QUASAR_NARROW_FORMATS,
-    DataFormat.Float16_b: _QUASAR_NARROW_FORMATS,
-    DataFormat.Int32: {DataFormat.Int32, DataFormat.Int8, DataFormat.UInt8},
-    DataFormat.Int16: {DataFormat.Int16},
-    DataFormat.Int8: {DataFormat.Int8},
-    DataFormat.UInt8: {DataFormat.UInt8},
-}
-
-
-def _quasar_sfpu_hardware_format(fmt: DataFormat) -> DataFormat:
-    """Map an SFPU-visible logical format to its Quasar hardware encoding."""
-    return DataFormat.Int16 if fmt == DataFormat.UInt16 else fmt
-
-
-def _quasar_dest_hardware_format(fmt: DataFormat) -> DataFormat:
-    if fmt == DataFormat.Tf32:
-        return DataFormat.Float32
-    return _quasar_sfpu_hardware_format(fmt)
-
-
-def _quasar_effective_sfpu_format(fmt: DataFormat) -> DataFormat:
-    """Return the register format exposed to SFPU after Quasar unpacking."""
-    if fmt.is_mx_format():
-        return DataFormat.Float16_b
-    if fmt == DataFormat.Fp8_e4m3:
-        return DataFormat.Float16
-    if fmt == DataFormat.Tf32:
-        return DataFormat.Float32
-    return _quasar_sfpu_hardware_format(fmt)
-
-
-def _quasar_fpu_source_format(fmt: DataFormat) -> DataFormat:
-    """Return the SrcA format consumed by the FPU datacopy staging path."""
-    if fmt in (DataFormat.Float32, DataFormat.Tf32):
-        return DataFormat.Tf32
-    if fmt.is_mx_format():
-        return DataFormat.Float16_b
-    if fmt == DataFormat.Fp8_e4m3:
-        return DataFormat.Float16
-    return _quasar_sfpu_hardware_format(fmt)
-
-
-def _quasar_dest_format_matches_mode(
-    dest_format: DataFormat, dest_acc: DestAccumulation
-) -> bool:
-    is_32_bit = dest_format in (DataFormat.Float32, DataFormat.Int32)
-    return is_32_bit == (dest_acc == DestAccumulation.Yes)
-
-
-def is_valid_quasar_unpack_to_dest(
-    input_format: DataFormat,
-    unpack_output_format: DataFormat,
-    dest_acc: DestAccumulation,
-    *,
-    allow_narrow_in_32bit_dest: bool = False,
-) -> bool:
-    """Validate an L1 -> Dest unpack conversion against the Tensix table.
-
-    Non-32bit -> 32bit data conversions are not supported by unpacker.
-    That layout is accepted only for SFPU typecast, which reads the
-    narrow representation and overwrites it with the converted result.
-    """
-    l1_input_format = _quasar_sfpu_hardware_format(input_format)
-    unpack_out_format = _quasar_dest_hardware_format(unpack_output_format)
-
-    # Check whether the unpacker supports the requested conversion.
-    if unpack_out_format not in _QUASAR_UNPACK_TO_DEST_FORMATS.get(
-        l1_input_format, set()
-    ):
-        return False
-
-    # Check that the register format matches the configured Dest width.
-    if _quasar_dest_format_matches_mode(unpack_out_format, dest_acc):
-        return True
-
-    # Typecast may temporarily hold a narrow input in a 32-bit Dest slot; SFPU
-    # overwrites it with the converted 32-bit result.
-    return allow_narrow_in_32bit_dest and dest_acc == DestAccumulation.Yes
-
-
-def is_valid_quasar_fpu_path(
-    input_format: DataFormat,
-    unpack_output_format: DataFormat,
-    dest_format: DataFormat,
-    dest_acc: DestAccumulation,
-) -> bool:
-    """Validate the complete Unpack-to-SrcA + FPU datacopy -> Dest path."""
-    # UInt16 uses the Int16 encoding for SFPU input/output, but Quasar does not
-    # support an Int16 Src -> UInt16 Dest FPU datacopy.
-    if dest_format == DataFormat.UInt16:
-        return False
-
-    # L1 input format.
-    hardware_input = _quasar_sfpu_hardware_format(input_format)
-    # SrcA/B format, also used as the math format.
-    hardware_unpack_output = _quasar_sfpu_hardware_format(unpack_output_format)
-    # Dest format, also used as the SFPU and packer input format.
-    hardware_dest = _quasar_sfpu_hardware_format(dest_format)
-
-    return (
-        hardware_unpack_output
-        in _QUASAR_UNPACK_TO_SRCA_FORMATS.get(hardware_input, set())
-        and hardware_dest
-        in _QUASAR_FPU_DATACOPY_DEST_FORMATS.get(hardware_unpack_output, set())
-        and _quasar_dest_format_matches_mode(hardware_dest, dest_acc)
-    )
-
-
-def is_valid_quasar_packer_conversion(
-    pack_input_format: DataFormat, output_format: DataFormat
-) -> bool:
-    """Validate a Dest -> L1 conversion against the Tensix packer table."""
-    hardware_input = _quasar_sfpu_hardware_format(pack_input_format)
-    hardware_output = _quasar_sfpu_hardware_format(output_format)
-    return hardware_output in _QUASAR_PACK_OUTPUTS.get(hardware_input, set())
 
 
 def _quasar_narrow_dest_candidates(
@@ -710,7 +526,6 @@ def _make_quasar_sfpu_variant(
     sfpu_src: DataFormat,
     sfpu_dst: DataFormat,
     pack_src: DataFormat,
-    int32_dest: bool = False,
 ) -> Optional[QuasarSfpuVariant]:
     if not is_valid_quasar_packer_conversion(pack_src, formats.output_format):
         return None
@@ -722,7 +537,6 @@ def _make_quasar_sfpu_variant(
         sfpu_src=sfpu_src,
         sfpu_dst=sfpu_dst,
         pack_src=pack_src,
-        int32_dest=int32_dest,
     )
 
 
@@ -804,10 +618,6 @@ def _resolve_quasar_fpu_sfpu_variant(
         sfpu_src=dest_format,
         sfpu_dst=dest_format,
         pack_src=dest_format,
-        int32_dest=(
-            fpu_source in (DataFormat.Int8, DataFormat.UInt8)
-            and dest_format == DataFormat.Int32
-        ),
     )
